fix ciruclar import and split data
